+++ neost/PosteriorAnalysis.py
@@ -2,13 +2,11 @@ import numpy
 import matplotlib
 from matplotlib.lines import Line2D
 import matplotlib.patches as mpatches
+from matplotlib.colors import ListedColormap
 from matplotlib import pyplot
 import seaborn as sns
-import getdist
-from getdist import plots
 from scipy.interpolate import UnivariateSpline
 from scipy.stats import gaussian_kde
-from alive_progress import alive_bar
 import corner
 
 import neost
@@ -61,6 +59,161 @@ def calc_bands(x, y):
     maxy = maxy[~numpy.all(maxy == 0, axis=1)]
     return miny, maxy
 
+def get_quantiles(array, quantiles=[0.025, 0.5, 0.975]):
+        contours = numpy.nanquantile(array, quantiles) #changed to nanquantile to inorder to ignore the nans that may appear
+        low = contours[0]
+        median = contours[1]
+        high = contours[2]
+        minus = low - median
+        plus = high - median
+        return numpy.round(median,2),numpy.round(plus,2),numpy.round(minus,2)  #returns uncertainties on the array
+
+def compute_table_data(root_name, EOS, variable_params, static_params):
+    ewposterior = numpy.loadtxt(root_name + 'post_equal_weights.dat')
+    print("Total number of samples is %d" %(len(ewposterior)))
+    try:
+        Data_array = numpy.loadtxt(root_name + 'table_data.txt')
+        print('M_TOV: ', get_quantiles(Data_array[:,0]))
+        print('R_TOV: ', get_quantiles(Data_array[:,1]))
+        print('eps_cent TOV: ', get_quantiles(Data_array[:,2]))
+        print('rho_cent TOV: ', get_quantiles(Data_array[:,3]))
+        print('P_cent TOV: ', get_quantiles(Data_array[:,4]))
+        print('R_1.4: ', get_quantiles(Data_array[:,5]))
+        print('eps_cent 1.4: ', get_quantiles(Data_array[:,6]))
+        print('rho_cent 1.4: ', get_quantiles(Data_array[:,7]))
+        print('P_cent 1.4: ', get_quantiles(Data_array[:,8]))
+        print('R_2.0: ', get_quantiles(Data_array[:,9]))
+        print('eps_cent 2.0: ', get_quantiles(Data_array[:,10]))
+        print('rho_cent 2.0: ', get_quantiles(Data_array[:,11]))
+        print('P_cent 2.0: ', get_quantiles(Data_array[:,12]))
+        print('Delta R = R_2.0 - R_1.4: ', get_quantiles(Data_array[:,9] - Data_array[:,5]))
+    except OSError:
+        Data_array = numpy.zeros((len(ewposterior),13)) #contains Mtov, Rtov, eps_cent TOV, rho_cent TOV, P_cent TOV,R 1.4, eps_cent 1.4, rho_cent 1.4,
+                                                                #P_cent 1.4, R 2.0, eps_cent 2.0, rho_cent 2.0, P_cent 2.0
+
+        for i in range(0, len(ewposterior), 1):
+            pr = ewposterior[i][0:len(variable_params)]
+            par = {e:pr[j] for j, e in enumerate(list(variable_params.keys()))}
+            par.update(static_params)
+            EOS.update(par, max_edsc=True)
+
+            edsrho = UnivariateSpline(EOS.energydensities, EOS.massdensities, k=1, s=0)
+            max_rhoc = edsrho(EOS.max_edsc) / rho_ns #division by rho_ns gives max_rhoc in terms of n_c/n_0 as mass density and number density only differ by a factor the mass of baryon, which is canceled out in this fraction
+
+            eps = numpy.logspace(14.4, numpy.log10(EOS.max_edsc), 40)
+            M = numpy.zeros(len(eps))
+            R = numpy.zeros(len(eps))
+            for j, e in enumerate(eps):
+                star = Star(e)
+                star.solve_structure(EOS.energydensities, EOS.pressures)
+                M[j] = star.Mrot
+                R[j] = star.Req
+
+            M, indices = numpy.unique(M, return_index=True)
+            MR = UnivariateSpline(M, R[indices], k=1, s=0, ext=1)
+            epsM = UnivariateSpline(M, eps[indices], k=1, s=0,ext = 1)
+
+            R_14 = MR(1.4)
+            if R_14 == 0:
+                R_14 = numpy.nan # set to be nan so they don't impact the quantiles b/c we are using numpy.nanquantiles
+                eps_14 = numpy,nan
+                rho_14 = numpy.nan
+                pres_14 = numpy.nan
+            else:
+                eps_14 = epsM(1.4)
+                rho_14 = edsrho(eps_14) / rho_ns
+                pres_14 = EOS.eos(eps_14)
+
+            R_2 = MR(2.0)
+            if R_2 == 0:
+                R_2 = numpy.nan # see above for reasoning
+                eps_2 = numpy.nan
+                rho_2 = numpy.nan
+                pres_2 = numpy.nan # see above for reasoning
+            else:
+                eps_2 = epsM(2.0)
+                rho_2 = edsrho(eps_2) / rho_ns
+                pres_2 = EOS.eos(eps_2)
+
+            #print(numpy.log10(EOS.max_edsc), numpy.log10(EOS.eos(EOS.max_edsc)),R_14, numpy.log10(eps_14), numpy.log10(pres_14),R_2, numpy.log10(eps_2), numpy.log10(pres_2))
+            Data_array[i] = EOS.max_M, EOS.Radius_max_M, numpy.log10(EOS.max_edsc), max_rhoc, numpy.log10(EOS.eos(EOS.max_edsc)),R_14, numpy.log10(eps_14), rho_14, numpy.log10(pres_14),R_2, numpy.log10(eps_2), rho_2, numpy.log10(pres_2)
+            if i%1000 == 0:
+                print(i)
+            #print(Data_array[i])
+    # save everything
+        numpy.savetxt(root_name + 'table_data.txt', Data_array)
+        print('M_TOV: ', get_quantiles(Data_array[:,0]))
+        print('R_TOV: ', get_quantiles(Data_array[:,1]))
+        print('eps_cent TOV: ', get_quantiles(Data_array[:,2]))
+        print('rho_cent TOV: ', get_quantiles(Data_array[:,3]))
+        print('P_cent TOV: ', get_quantiles(Data_array[:,4]))
+        print('R_1.4: ', get_quantiles(Data_array[:,5]))
+        print('eps_cent 1.4: ', get_quantiles(Data_array[:,6]))
+        print('rho_cent 1.4: ', get_quantiles(Data_array[:,7]))
+        print('P_cent 1.4: ', get_quantiles(Data_array[:,8]))
+        print('R_2.0: ', get_quantiles(Data_array[:,9]))
+        print('eps_cent 2.0: ', get_quantiles(Data_array[:,10]))
+        print('rho_cent 2.0: ', get_quantiles(Data_array[:,11]))
+        print('P_cent 2.0: ', get_quantiles(Data_array[:,12]))
+        print('Delta R = R_2.0 - R_1.4: ', get_quantiles(Data_array[:,9] - Data_array[:,5]))
+
+def compute_prior_auxiliary_data(root_name, EOS, variable_params, static_params):
+    ewprior = numpy.loadtxt(root_name + 'post_equal_weights.dat')
+    print("Total number of samples is %d" %(len(ewprior)))
+
+    num_stars = len(numpy.array([v for k,v in variable_params.items() if 'rhoc' in k]))
+
+    if len(list(variable_params.keys())) == num_stars:
+        flag = True
+
+    else:
+        flag = False
+
+    masses = numpy.linspace(.2, 2.9, 50)
+    energydensities = numpy.logspace(14.2, 16, 50)
+
+    if flag == True:
+        pressures = numpy.zeros((len(masses), len(ewprior)))
+        pressures_rho = numpy.zeros((len(masses), len(ewprior)))
+        MR_prpr_pp = numpy.zeros((len(ewprior), 2))
+    else:
+        pressures = numpy.zeros((len(masses), len(ewprior)))
+        pressures_rho = numpy.zeros((len(masses), len(ewprior)))
+        minpres = numpy.zeros((3, len(energydensities)))
+        maxpres = numpy.zeros((3, len(energydensities)))
+        minpres_rho = numpy.zeros((3, len(energydensities)))
+        maxpres_rho = numpy.zeros((3, len(energydensities)))
+        MR_prpr_pp = numpy.zeros((len(ewprior), 2))
+
+    for i in range(0, len(ewprior), 1):
+
+        pr = ewprior[i][0:len(variable_params)]
+        par = {e:pr[j] for j, e in enumerate(list(variable_params.keys()))}
+        par.update(static_params)
+        EOS.update(par, max_edsc=True)
+
+        rhopres = UnivariateSpline(EOS.massdensities, EOS.pressures, k=1, s=0)
+        edsrho = UnivariateSpline(EOS.energydensities, EOS.massdensities, k=1, s=0)
+        max_rhoc = edsrho(EOS.max_edsc)
+        pressures_rho[:,i][energydensities<max_rhoc] = rhopres(energydensities[energydensities<max_rhoc])
+        pressures[:,i][energydensities<EOS.max_edsc] = EOS.eos(energydensities[energydensities<EOS.max_edsc])
+
+        rhoc = 10**par['rhoc_1'] #just pick one of the central density samples, their distributions will be identical since constant likelihood eval on all sources
+        star = Star(rhoc)
+        star.solve_structure(EOS.energydensities, EOS.pressures)
+        MR_prpr_pp[i] = star.Mrot, star.Req
+    # save everything
+    numpy.save(root_name + 'pressures', pressures)
+    numpy.savetxt(root_name + 'MR_prpr.txt', MR_prpr_pp)
+
+    if flag == False:
+        minpres, maxpres = calc_bands(energydensities, pressures)
+        minpres_rho, maxpres_rho = calc_bands(energydensities, pressures_rho)
+        numpy.save(root_name + 'minpres_rho', minpres_rho)
+        numpy.save(root_name + 'maxpres_rho', maxpres_rho)
+        numpy.save(root_name + 'minpres', minpres)
+        numpy.save(root_name + 'maxpres', maxpres)
+
 
 def compute_auxiliary_data(root_name, EOS, variable_params, static_params, chirp_masses): 
     ewposterior = numpy.loadtxt(root_name + 'post_equal_weights.dat')
@@ -100,55 +253,54 @@ def compute_auxiliary_data(root_name, EOS, variable_params, static_params, chirp
     
 
 
-    with alive_bar(len(ewposterior)) as bar:
-        for i in range(0, len(ewposterior), 1):
+    for i in range(0, len(ewposterior), 1):
 
-            pr = ewposterior[i][0:len(variable_params)]
-            par = {e:pr[j] for j, e in enumerate(list(variable_params.keys()))}
-            par.update(static_params)
-            EOS.update(par, max_edsc=True)
+        pr = ewposterior[i][0:len(variable_params)]
+        par = {e:pr[j] for j, e in enumerate(list(variable_params.keys()))}
+        par.update(static_params)
+        EOS.update(par, max_edsc=True)
 
-            rhopres = UnivariateSpline(EOS.massdensities, EOS.pressures, k=1, s=0)
-            edsrho = UnivariateSpline(EOS.energydensities, EOS.massdensities, k=1, s=0)
-            max_rhoc = edsrho(EOS.max_edsc)
-            pressures_rho[:,i][energydensities<max_rhoc] = rhopres(energydensities[energydensities<max_rhoc])
-            pressures[:,i][energydensities<EOS.max_edsc] = EOS.eos(energydensities[energydensities<EOS.max_edsc])
+        rhopres = UnivariateSpline(EOS.massdensities, EOS.pressures, k=1, s=0)
+        edsrho = UnivariateSpline(EOS.energydensities, EOS.massdensities, k=1, s=0)
+        max_rhoc = edsrho(EOS.max_edsc)
+        pressures_rho[:,i][energydensities<max_rhoc] = rhopres(energydensities[energydensities<max_rhoc])
+        pressures[:,i][energydensities<EOS.max_edsc] = EOS.eos(energydensities[energydensities<EOS.max_edsc])
             
-            rhocs = numpy.logspace(14.5, numpy.log10(EOS.max_edsc), 30)
-            M = numpy.zeros(len(rhocs))
-            R = numpy.zeros(len(rhocs))
-            for j, e in enumerate(rhocs):
-                star = Star(e)
-                star.solve_structure(EOS.energydensities, EOS.pressures)
-                M[j] = star.Mrot
-                R[j] = star.Req
-
-            M, indices = numpy.unique(M, return_index=True)
-            MR = UnivariateSpline(M, R[indices], k=1, s=0, ext=1)
-            rhocM = UnivariateSpline(M, rhocs[indices], k=1, s=0)
-            
-            rhocpar = numpy.array([10**v for k,v in par.items() if 'rhoc' in k])
-            tmp = []
-            for j, e in enumerate(rhocpar):
-                star = Star(e)
-                star.solve_structure(EOS.energydensities, EOS.pressures)
-                tmp.append([e, EOS.eos(e), star.Mrot, star.Req, star.tidal])
-
-                if chirp_masses[j] is not None: 
-                    M2 = m1(chirp_masses[j], tmp[j][2])
-                    rhoc = rhocM(M2)
-                    star = Star(rhoc)
-                    star.solve_structure(EOS.energydensities, EOS.pressures)
-                    tmp.append([rhoc, EOS.eos(rhoc), star.Mrot, star.Req, star.tidal])
-
-            scattered.append(tmp)
-            rhoc = numpy.random.rand() *(numpy.log10(EOS.max_edsc) - 14.6) + 14.6
-            star = Star(10**rhoc)
+        rhocs = numpy.logspace(14.5, numpy.log10(EOS.max_edsc), 30)
+        M = numpy.zeros(len(rhocs))
+        R = numpy.zeros(len(rhocs))
+        for j, e in enumerate(rhocs):
+            star = Star(e)
             star.solve_structure(EOS.energydensities, EOS.pressures)
-            MR_prpr_pp[i] = star.Mrot, star.Req
+            M[j] = star.Mrot
+            R[j] = star.Req
 
-            radii[:,i] = MR(masses)
-            bar()
+        M, indices = numpy.unique(M, return_index=True)
+        MR = UnivariateSpline(M, R[indices], k=1, s=0, ext=1)
+        rhocM = UnivariateSpline(M, rhocs[indices], k=1, s=0)
+            
+        rhocpar = numpy.array([10**v for k,v in par.items() if 'rhoc' in k])
+        tmp = []
+        for j, e in enumerate(rhocpar):
+            star = Star(e)
+            star.solve_structure(EOS.energydensities, EOS.pressures)
+            tmp.append([e, EOS.eos(e), star.Mrot, star.Req, star.tidal])
+
+            if chirp_masses[j] is not None:
+                M2 = m1(chirp_masses[j], tmp[j][2])
+                rhoc = rhocM(M2)
+                star = Star(rhoc)
+                star.solve_structure(EOS.energydensities, EOS.pressures)
+                tmp.append([rhoc, EOS.eos(rhoc), star.Mrot, star.Req, star.tidal])
+
+        scattered.append(tmp)
+        rhoc = numpy.random.rand() *(numpy.log10(EOS.max_edsc) - 14.6) + 14.6
+        star = Star(10**rhoc)
+        star.solve_structure(EOS.energydensities, EOS.pressures)
+        MR_prpr_pp[i] = star.Mrot, star.Req
+
+        radii[:,i] = MR(masses)
+
 
     scattered = numpy.array(scattered)
     # save everything
@@ -180,7 +332,7 @@ def mass_radius_posterior_plot(root_name):
     figure, ax = pyplot.subplots(1,1, figsize=(9,6))
     M_max = 0.
     for i in range(len(scatter[0])):
-        corner.hist2d(scatter[:,i][:,3], scatter[:,i][:,2], labels = ['R [km]', 'M [M$_{\odot}$]'], show_titles=True, 
+        corner.hist2d(scatter[:,i][:,3], scatter[:,i][:,2], labels = ['R [km]', r'M [M$_{\odot}$]'], show_titles=True,
                         color=colors[i], smooth=.8, data_kwargs={'ms':5, 'alpha':0.5})
         M_max = max(max(scatter[:,i][:,2]), M_max)
 
@@ -193,7 +345,7 @@ def mass_radius_posterior_plot(root_name):
     pyplot.show()
     figure.savefig(root_name + 'MRposterior.png')
 
-def mass_radius_prior_predictive_plot(root_name,variable_params, label_name='updated prior', prior_mrpredictive=None):
+def mass_radius_prior_predictive_plot(root_name,variable_params, label_name='updated prior'):
     fig, ax = pyplot.subplots(1,1, figsize=(9, 6))
 
     num_stars = len(numpy.array([v for k,v in variable_params.items() if 'rhoc' in k]))
@@ -207,21 +359,12 @@ def mass_radius_prior_predictive_plot(root_name,variable_params, label_name='upd
     if flag == True:
         raise Exception("Cannot perform mass_radius_prior_predictive_plot function because EoS is fixed, i.e., tabulated or all EoS params are static params!")
     else:
-        if prior_mrpredictive is not None:
-            MRprior = numpy.load(prior_mrpredictive) 
-            inbins = numpy.histogramdd(MRprior, bins=50, normed=True)
-            levels = getdist.densities.getContourLevels(inbins[0], contours=[0.95])
-            sns.kdeplot(x= MRprior[:,1], y=  MRprior[:,0], gridsize=50, 
-                        shade=False, ax=ax, n_levels=levels[::-1], linewidths=2,
-                        alpha=1., cmap=None, colors='black', linestyles='--')
-
         MR_prpr= numpy.loadtxt(root_name + 'MR_prpr.txt')
-        inbins = numpy.histogramdd(MR_prpr[:,[1,0]], bins=50, normed=True)
-        levels = getdist.densities.getContourLevels(inbins[0], contours=[0.68, 0.95])
+        inbins = numpy.histogramdd(MR_prpr[:,[1,0]], bins=50, density=True)
+        levels = [0.05, 0.32, 1]
     
-        sns.kdeplot(x = MR_prpr[:,1], y = MR_prpr[:,0], gridsize=50, 
-                    shade=True, ax=ax, n_levels=numpy.array([levels[1],levels[0], 1.]),
-                    alpha=1., cmap=None, colors=colors[[5,3]])
+        sns.kdeplot(x=MR_prpr[:,1], y=MR_prpr[:,0], gridsize=50, fill=True, ax=ax, levels=levels,
+                    alpha=1., cmap=ListedColormap(colors[[5,3]]))
 
         # add legend, for now 'prior' is always shown
         custom_lines = [Line2D([0], [0], color='black', lw=2, linestyle='--', alpha=1.),