+++ neost/__init__.py
@@ -1,7 +1,7 @@
 """ NEoST: Open-source code for equation of state inference via nested sampling. """
 from __future__ import print_function
-__version__ = "0.9.0"
-__author__ = "Geert Raaijmakers"
+__version__ = "0.10.0"
+__author__ = "The NEoST core team"
 
 try:
     __NEOST_SETUP__